Update transmittance +  I FOUND MY WRONG NUMBER
so you know how like when you want to do like phase functions, they gotta be all like kosher, like integrate to one, yaknow? so for rayleigh its like 3/16PI, which is great and all, but for some reason I had accidentally written PI*3/16 in my code so it was ALL WRONG and I was always so confused as to why mie and ambient and literally everything else was so dark compared to rayleigh and IT WAS BECAUSE I WAS GIVING IT PI^2 MORE WEIGHT!!! literally an order of magnitude. Never would have found it if not for this overhaul so yea thanks life, I don't know whether to be happy or depressed.